+++ adapters/DTExchange/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 #### Next version
 - Added bidding banner.
+- Added bidding interstitial.
 
 #### [Version 8.3.6.0](https://dl.google.com/googleadmobadssdk/mediation/ios/dtexchange/DTExchangeAdapter-8.3.6.0.zip)
 - Enabled `-fobjc-arc` and `-fstack-protector-all` flags.